mvc layout page ve sayfa tasarımları hazırlandı. ürünler listelendi. dependency injection ayağa kaldırıldı

+++ Northwind/Norhwind.Entities/Product.cs
@@ -10,7 +10,7 @@ namespace Norhwind.Entities
     {
         public int ProductID { get; set; }
         public string ProductName { get; set; }
-        public int UnitsInStock { get; set; }
+        public short UnitsInStock { get; set; }
         public decimal UnitPrice { get; set; }
         public int CategoryID { get; set; }
     }
